+++ scheduled-actions.tf
@@ -1,28 +1,15 @@
-resource "aws_appautoscaling_scheduled_action" "lights_on" {
-  count              = length(var.lights_on_schedule_expr) == 0 ? 0 : 1
-  name               = "${var.family}-lights-on"
+resource "aws_appautoscaling_scheduled_action" "actions" {
+  count              = length(var.scheduled_actions)
+  name               = "${var.family}-scheduled-action-${count.index}"
   service_namespace  = aws_appautoscaling_target.ecs_target.service_namespace
   resource_id        = aws_appautoscaling_target.ecs_target.resource_id
   scalable_dimension = aws_appautoscaling_target.ecs_target.scalable_dimension
-  schedule           = var.lights_on_schedule_expr
-  timezone           = var.schedule_timezone
+  schedule           = var.scheduled_actions[count.index].expression
+  timezone           = var.scheduled_actions_timezone
 
   scalable_target_action {
-    min_capacity = var.min_capacity >= 0 ? var.min_capacity : var.desired_capacity
-    max_capacity = var.max_capacity >= 0 ? var.max_capacity : var.desired_capacity
+    min_capacity = var.scheduled_actions[count.index].min_capacity
+    max_capacity = var.scheduled_actions[count.index].max_capacity
   }
 }
-resource "aws_appautoscaling_scheduled_action" "lights_off" {
-  count              = length(var.lights_off_schedule_expr) == 0 ? 0 : 1
-  name               = "${var.family}-lights-off"
-  service_namespace  = aws_appautoscaling_target.ecs_target.service_namespace
-  resource_id        = aws_appautoscaling_target.ecs_target.resource_id
-  scalable_dimension = aws_appautoscaling_target.ecs_target.scalable_dimension
-  schedule           = var.lights_off_schedule_expr
-  timezone           = var.schedule_timezone
 
-  scalable_target_action {
-    min_capacity = var.lights_off_desired_capacity
-    max_capacity = var.lights_off_desired_capacity
-  }
-}